server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Biswajit Patra <patrabiswajit133@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
gunicorn = "^20.1.0"
dj-database-url = "^0.5.0"
whitenoise = "^5.2.0"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.4"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
